requirements.txt
transformers
torch
matplotlib
numpy
Pillow
opencv-python
opencv-contrib-python